requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
sentence-transformers
pydantic
